backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scriptorium-ai-backend"
version = "0.1.0"
description = "Backend Scriptorium AI — plateforme générique d'éditions savantes augmentées"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "google-genai>=1.0",
    "google-auth>=2.0",
    "httpx>=0.27",
    "lxml>=5.2",
    "Pillow>=10.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
